Fixed type issue for LuaScriptable giveItem() export
git-svn-id: svn://70.113.44.214/VanaMirror/trunk@1515 09c1253f-3d09-8e40-9719-4320e56601b5

+++ src/ChannelServer/LuaScriptable.cpp
@@ -168,7 +168,7 @@ int LuaExports::addSkillLevel(lua_State *luaVm) {
 
 int LuaExports::giveItem(lua_State *luaVm) {
 	int32_t itemid = lua_tointeger(luaVm, -2);
-	int8_t amount = lua_tointeger(luaVm, -1);
+	int16_t amount = lua_tointeger(luaVm, -1);
 	bool success = Quests::giveItem(getPlayer(luaVm), itemid, amount);
 
 	lua_pushboolean(luaVm, success);
